+++ BioDude/Assets/Scenes/initial.unity
@@ -6542,7 +6542,6 @@ GameObject:
   - component: {fileID: 1753083373}
   - component: {fileID: 1753083372}
   - component: {fileID: 1753083371}
-  - component: {fileID: 1753083374}
   - component: {fileID: 1753083375}
   m_Layer: 8
   m_Name: WallMap
@@ -6572,7 +6571,7 @@ TilemapCollider2D:
   m_GameObject: {fileID: 1753083369}
   m_Enabled: 1
   m_Density: 1
-  m_Material: {fileID: 6200000, guid: 5961403e8f2d04e4794e4eba5168654c, type: 2}
+  m_Material: {fileID: 0}
   m_IsTrigger: 0
   m_UsedByEffector: 0
   m_UsedByComposite: 0
@@ -6862,22 +6861,6 @@ Tilemap:
     m_DirtyIndex: 0
   - serializedVersion: 1
     m_DirtyIndex: 0
---- !u!114 &1753083374
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 1753083369}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f0d8271b9b617c143bd412eb350e0b1b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  currDir: {x: 0, y: 0, z: 0}
-  collisionMask:
-    serializedVersion: 2
-    m_Bits: 256
-  maxStepDistance: 200
 --- !u!50 &1753083375
 Rigidbody2D:
   serializedVersion: 4
@@ -6891,12 +6874,12 @@ Rigidbody2D:
   m_UseAutoMass: 1
   m_Mass: 30.071299
   m_LinearDrag: 0
-  m_AngularDrag: 1
+  m_AngularDrag: 1000000
   m_GravityScale: 1
   m_Material: {fileID: 0}
   m_Interpolate: 0
   m_SleepingMode: 1
-  m_CollisionDetection: 0
+  m_CollisionDetection: 1
   m_Constraints: 7
 --- !u!1 &1755524264
 GameObject:
