+++ BioDude/Assets/Scripts/Items scripts/Weapon Scripts/Richochet.cs
@@ -1,0 +1,54 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Richochet : MonoBehaviour {
+
+
+    private LayerMask wallMask;
+    private Rigidbody2D rigidbody;
+
+
+    public void Awake()
+    {
+        wallMask = LayerMask.GetMask("Collider");
+        rigidbody = transform.GetComponent<Rigidbody2D>();
+    }
+
+    // Use this for initialization
+    void Start () {
+		
+	}
+	
+	// Update is called once per frame
+	void Update () {
+		
+	}
+    private Vector3 oldVelocity;
+    void FixedUpdate()
+    {
+        // because we want the velocity after physics, we put this in fixed update
+        oldVelocity = rigidbody.velocity;
+    }
+
+    public void OnCollisionEnter2D(Collision2D collision)
+    {
+        Debug.Log(collision.transform.tag);
+        if (collision.transform.tag == "Bouncy")
+        {
+            Debug.Log("hit");
+            // get the point of contact
+            ContactPoint2D contact = collision.contacts[0];
+
+            // reflect our old velocity off the contact point's normal vector
+            Vector3 reflectedVelocity = Vector3.Reflect(oldVelocity, contact.normal);
+
+            // assign the reflected velocity back to the rigidbody
+            rigidbody.velocity = reflectedVelocity;
+            // rotate the object by the same ammount we changed its velocity
+            Quaternion rotation = Quaternion.FromToRotation(oldVelocity, reflectedVelocity);
+            transform.rotation = rotation * transform.rotation;
+        }
+        
+    }
+}
