+++ BioDude/Assets/Scripts/Items scripts/Weapon Scripts/Richochet.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: f0d8271b9b617c143bd412eb350e0b1b
-timeCreated: 1525764997
+guid: 8144356e17304b445adae331135f2e6c
+timeCreated: 1526921195
 licenseType: Free
 MonoImporter:
   externalObjects: {}
