--- BioDude/Assets/Scripts/bouncy wall.physicsMaterial2D
@@ -1,10 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!62 &6200000
-PhysicsMaterial2D:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_Name: bouncy wall
-  friction: 0
-  bounciness: 1
