--- BioDude/Assets/Scripts/bouncy wall.physicsMaterial2D.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: 5961403e8f2d04e4794e4eba5168654c
-timeCreated: 1526313659
-licenseType: Free
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 6200000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
